+++ docs/changes.md
@@ -16,6 +16,7 @@ The stable 1.0.0 release freezes the API of funcparserlib 0.3.6 which was releas
 * Added support for Python 3.8
 * Added `-p` (the same as `skip(p)`) with more strict type hints for `-p` and `p1 + p2`
 * Added `tok(type[, value])` for more compact grammars, better error messages
+* Added `TokenSpec(type, pattern[, flags])` to simplify the use of `make_tokenizer()`
 * Added type hints for the public API
 
 ### Changed